+++ packages/jest-reporters/src/SummaryReporter.ts
@@ -19,8 +19,6 @@ import getSnapshotSummary from './getSnapshotSummary';
 import getSummary from './getSummary';
 import type {ReporterOnStartOptions, SummaryReporterOptions} from './types';
 
-const TEST_SUMMARY_THRESHOLD = 20;
-
 const NPM_EVENTS = new Set([
   'prepublish',
   'publish',
@@ -54,7 +52,7 @@ const {npm_config_user_agent, npm_lifecycle_event, npm_lifecycle_script} =
 export default class SummaryReporter extends BaseReporter {
   private _estimatedTime: number;
   private readonly _globalConfig: Config.GlobalConfig;
-  private readonly _options: SummaryReporterOptions;
+  private readonly _summaryThreshold: number;
 
   static readonly filename = __filename;
 
@@ -64,8 +62,18 @@ export default class SummaryReporter extends BaseReporter {
   ) {
     super();
     this._globalConfig = globalConfig;
-    this._options = options;
     this._estimatedTime = 0;
+    this.validateOptions(options);
+    this._summaryThreshold = options.summaryThreshold || 20;
+  }
+
+  private validateOptions(options: SummaryReporterOptions) {
+    if (
+      options.summaryThreshold &&
+      typeof options.summaryThreshold !== 'number'
+    ) {
+      throw new TypeError('The option summaryThreshold should be a number');
+    }
   }
 
   // If we write more than one character at a time it is possible that
@@ -181,8 +189,7 @@ export default class SummaryReporter extends BaseReporter {
     const runtimeErrors = aggregatedResults.numRuntimeErrorTestSuites;
     if (
       failedTests + runtimeErrors > 0 &&
-      (this._options.forceFailSummary ||
-        aggregatedResults.numTotalTestSuites > TEST_SUMMARY_THRESHOLD)
+      aggregatedResults.numTotalTestSuites > this._summaryThreshold
     ) {
       this.log(chalk.bold('Summary of all failing tests'));
       aggregatedResults.testResults.forEach(testResult => {